+++ bitwarden_license/bit-cli/webpack.config.js
@@ -1,10 +1,12 @@
-const path = require("path");
-const { buildConfig } = require("../../apps/cli/webpack.base");
+const TsconfigPathsPlugin = require("tsconfig-paths-webpack-plugin");
 
-module.exports = buildConfig({
-  configName: "Commercial",
-  entry: {
-    bw: path.resolve(__dirname, "src/bw.ts"),
-  },
-  tsConfig: path.resolve(__dirname, "tsconfig.json"),
-});
+// Re-use the OSS CLI webpack config
+const webpackConfig = require("../../apps/cli/webpack.config");
+
+// Update paths to use the bit-cli entrypoint and tsconfig
+webpackConfig.entry = { bw: "../../bitwarden_license/bit-cli/src/bw.ts" };
+webpackConfig.resolve.plugins = [
+  new TsconfigPathsPlugin({ configFile: "../../bitwarden_license/bit-cli/tsconfig.json" }),
+];
+
+module.exports = webpackConfig;
